backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
elevenlabs==0.2.27
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

